src/requirements.txt
numpy~=1.20.1
unidecode~=1.3.2
lxml~=4.6.3
pandas~=1.2.4
matplotlib~=3.5.0
setuptools>=52.0.0
sqlalchemy>=1.4.7
python-dotenv>=0.19.2
fiona~=1.8.20
shapely~=1.7.1
tqdm~=4.64.0
scipy~=1.6.2